+++ packages/i18n/index.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <title>Tiny Vue WebComponent</title>
+    <title>Tiny Vue i18n</title>
   </head>
 
   <body>
