+++ packages/i18n/package.json
@@ -28,8 +28,6 @@
   "homepage": "https://opentiny.design/tiny-engine",
   "devDependencies": {
     "@vitejs/plugin-vue": "^5.1.2",
-    "eslint": "^7.32.0",
-    "prettier": "^2.4.0",
     "vite": "^5.4.2"
   },
   "peerDependencies": {
