--- packages/i18n/src/test/TestWebcomponent.vue
@@ -1,20 +1,0 @@
-<template>
-  <h4>webcomponent</h4>
-  <p>{{ t('hello') }}</p>
-  <hr />
-</template>
-
-<script lang="ts">
-import { inject } from 'vue'
-import { I18nInjectionKey } from 'vue-i18n'
-
-export default {
-  setup() {
-    const {
-      global: { t }
-    } = inject<any>(I18nInjectionKey)
-
-    return { t }
-  }
-}
-</script>
